third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-sizes-js-change-expected.txt

main frame - img requested 'image-set-4x.png'
PASS successfullyParsed is true

TEST COMPLETE
PASS currentSrcFileName("foo") is "image-set-4x.png"
